Title: 소괄호 매칭 수 최대화하기, Time: 638ms, Memory: 29MB, Status: Wrong Answer - Codetree

+++ 250329/소괄호 매칭 수 최대화하기/maximize-the-number-of-parenthesis-matches.java
@@ -24,7 +24,7 @@ class Pair implements Comparable<Pair> {
 public class Main {
 
     public static Pair[] pairs;
-    public static int score;
+    public static long score;
     public static long allR;
 
     public static void main(String[] args) {
